fixed bugs in email formating
