+++ WaterLog_Backend/WaterLog_Backend.csproj
@@ -15,4 +15,8 @@
     <PackageReference Include="System.Net.Mail" Version="1.0.0-rtm-00002" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\EmailNotifications\EmailNotifications.csproj" />
+  </ItemGroup>
+
 </Project>
